-Wsign-conversion and -Wstrict-prototypes

+++ src/main.c
@@ -73,10 +73,10 @@ int main(int argc, char **argv)
             const int depth = argc > 2 ? atoi(argv[2]) : 12;
 
             if (argc > 3)
-                WorkersCount = atoi(argv[3]);
+                WorkersCount = (size_t)atoll(argv[3]);
 
             if (argc > 4)
-                uciHash = 1ULL << bb_msb(atoll(argv[4]));  // must be a power of 2
+                uciHash = 1ULL << bb_msb((uint64_t)atoll(argv[4]));  // must be a power of 2
 
             workers_prepare(WorkersCount);
             hash_prepare(uciHash);
